requirements.txt
asn1crypto==0.22.0
astroid==1.5.3
aws==0.2.5
awscli==1.11.141
bcrypt==3.1.3
boto3==1.4.6
botocore==1.6.8
cffi==1.10.0
colorama==0.3.7
coverage==4.4.1
cryptography==2.0.3
Django==1.11.3
django-cors-headers==2.1.0
django-cors-middleware==1.3.1
django-coverage==1.2.4
docutils==0.14
Fabric==1.13.2
idna==2.6
isort==4.2.15
jmespath==0.9.3
lazy-object-proxy==1.3.1
mccabe==0.6.1
mock==2.0.0
olefile==0.44
paramiko==2.2.1
pbr==3.1.1
Pillow==4.2.1
prettytable==0.7.2
psycopg2==2.7.1
pyasn1==0.3.2
pycparser==2.18
PyJWT==1.5.3
pylint==1.7.2
pylint-django==0.7.2
pylint-plugin-utils==0.2.6
PyNaCl==1.1.2
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
rsa==3.4.2
s3transfer==0.1.10
six==1.10.0
wrapt==1.10.10
